Remake

+++ src/main/java/ru/netology/model/Post.java
@@ -1,5 +1,6 @@
 package ru.netology.model;
 
+
 public class Post {
   private long id;
   private String content;
@@ -16,7 +17,7 @@ public class Post {
     return id;
   }
 
-  public void setId(long id) {
+  public void setId(Integer id) {
     this.id = id;
   }
